+++ core/src/main/java/io/github/jimzhouzzy/klotski/logic/KlotskiSolver.java
@@ -1,4 +1,4 @@
-package io.github.jimzhouzzy.klotski;
+package io.github.jimzhouzzy.klotski.logic;
 
 import java.util.ArrayDeque;
 import java.util.ArrayList;
