+++ core/src/main/java/io/github/jimzhouzzy/klotski/ui/KlotskiTheme.java
@@ -1,0 +1,6 @@
+package io.github.jimzhouzzy.klotski.ui;
+
+public enum KlotskiTheme {
+    DARK,
+    LIGHT
+}
